requirements-vercel.txt
# Core AI/LLM (Stateless version)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-mistralai>=0.2.0
mistralai>=1.0.0

# Vector Database (Cloud)
langchain-pinecone>=0.2.0
pinecone-client>=5.0.0

# Document Loaders (Lightweight)
pypdf>=4.0.0

# API Layer
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12

# Config & Utils
pydantic>=2.7.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Type hints
typing-extensions>=4.12.0
